Get rid of rug depenedency
