+++ .github/workflows/rust.yml
@@ -57,9 +57,6 @@ jobs:
         run: cargo build --benches --features experimental,copy_key,unsecure_schemes
       - name: cargo test
         run: cargo nextest run --features experimental,copy_key,unsecure_schemes
-      - name: Test VDF with GMP
-        if: matrix.os == 'ubuntu-ghcloud'
-        run: cd fastcrypto-vdf && cargo nextest run --features gmp && cd ..
       - name: Doctests
         run: |
           cargo test --doc --features experimental,copy_key,unsecure_schemes
