+++ fastcrypto-cli/Cargo.toml
@@ -8,9 +8,6 @@ publish = ["crates-io"]
 description = "Command line tools for cryptographic functions"
 repository = "https://github.com/MystenLabs/fastcrypto"
 
-[features]
-gmp = ["fastcrypto-vdf/gmp"]
-
 [dependencies]
 clap = { version = "4.1.8", features = ["derive"] }
 fastcrypto = { path = "../fastcrypto", features = ["copy_key"] }
